+++ fluent-dom/CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
 ## Unreleased
-  - …
+
+  - Drop support for IE and old evergreen browsers. (#133)
+
+    Currently supported are: Firefox 52+, Chrome 55+, Edge 15+, Safari 10.1+,
+    iOS Safari 10.3+ and node 8.9+.
 
 ## fluent-dom 0.3.0
+
   - Refactor the overlay sanitization methods into separate functions. (#189)
   - Separate out CachedIterable and CachedAsyncIterable, and add a param to touchNext. (#191)
   - Localization.formatValues should accept an array of objects. (#198)