+++ fluent-langneg/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
-  - …
+  - Drop support for IE and old evergreen browsers. (#133)
+
+    Currently supported are: Firefox 52+, Chrome 55+, Edge 15+, Safari 10.1+,
+    iOS Safari 10.3+ and node 8.9+.
 
 ## fluent-langneg 0.1.0 (August 17, 2017)