+++ fluent/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+  - Drop support for IE and old evergreen browsers. (#133)
+
+    Currently supported are: Firefox 52+, Chrome 55+, Edge 15+, Safari 10.1+,
+    iOS Safari 10.3+ and node 8.9+.
+
 ## fluent 0.6.4 (April 11, 2018)
 
   - Minor optimization to bidirectionality isolation